+++ Hengband/Hengband/Hengband.vcxproj
@@ -418,10 +418,10 @@
     <ClCompile Include="..\..\src\player-info\resistance-info.cpp" />
     <ClCompile Include="..\..\src\player-info\self-info-util.cpp" />
     <ClCompile Include="..\..\src\player-info\weapon-effect-info.cpp" />
-    <ClCompile Include="..\..\src\player-ability\player-speed.cpp" />
-    <ClCompile Include="..\..\src\player-ability\player-status-base.cpp" />
-    <ClCompile Include="..\..\src\player-ability\player-stealth.cpp" />
-    <ClCompile Include="..\..\src\player-ability\player-basic-statistics.cpp" />
+    <ClCompile Include="..\..\src\player-status\player-speed.cpp" />
+    <ClCompile Include="..\..\src\player-status\player-status-base.cpp" />
+    <ClCompile Include="..\..\src\player-status\player-stealth.cpp" />
+    <ClCompile Include="..\..\src\player-status\player-basic-statistics.cpp" />
     <ClCompile Include="..\..\src\player-ability\player-strength.cpp" />
     <ClCompile Include="..\..\src\player-ability\player-intelligence.cpp" />
     <ClCompile Include="..\..\src\player-ability\player-wisdom.cpp" />
@@ -1108,10 +1108,10 @@
     <ClInclude Include="..\..\src\player-info\resistance-info.h" />
     <ClInclude Include="..\..\src\player-info\self-info-util.h" />
     <ClInclude Include="..\..\src\player-info\weapon-effect-info.h" />
-    <ClInclude Include="..\..\src\player-ability\player-speed.h" />
-    <ClInclude Include="..\..\src\player-ability\player-status-base.h" />
-    <ClInclude Include="..\..\src\player-ability\player-stealth.h" />
-    <ClInclude Include="..\..\src\player-ability\player-basic-statistics.h" />
+    <ClInclude Include="..\..\src\player-status\player-speed.h" />
+    <ClInclude Include="..\..\src\player-status\player-status-base.h" />
+    <ClInclude Include="..\..\src\player-status\player-stealth.h" />
+    <ClInclude Include="..\..\src\player-status\player-basic-statistics.h" />
     <ClInclude Include="..\..\src\player-ability\player-strength.h" />
     <ClInclude Include="..\..\src\player-ability\player-intelligence.h" />
     <ClInclude Include="..\..\src\player-ability\player-wisdom.h" />
