+++ src/mind/mind-elementalist.cpp
@@ -36,8 +36,8 @@
 #include "monster-race/race-flags3.h"
 #include "monster-race/race-flags7.h"
 #include "monster/monster-describer.h"
-#include "player-ability/player-status-base.h"
 #include "player-info/avatar.h"
+#include "player-status/player-status-base.h"
 #include "player/player-status-table.h"
 #include "player/player-status.h"
 #include "racial/racial-util.h"
