+++ src/player-ability/player-charisma.h
@@ -1,5 +1,5 @@
 ﻿#pragma once
-#include "player-ability/player-basic-statistics.h"
+#include "player-status/player-basic-statistics.h"
 
 class PlayerCharisma : public PlayerBasicStatistics {
 public:
