+++ src/player-ability/player-constitution.h
@@ -1,5 +1,5 @@
 ﻿#pragma once
-#include "player-ability/player-basic-statistics.h"
+#include "player-status/player-basic-statistics.h"
 
 class PlayerConstitution : public PlayerBasicStatistics {
 public:
