+++ src/player/player-status-flags.cpp
@@ -13,9 +13,9 @@
 #include "object-hook/hook-checker.h"
 #include "object-hook/hook-weapon.h"
 #include "object/object-flags.h"
-#include "player-ability/player-basic-statistics.h"
-#include "player-ability/player-speed.h"
-#include "player-ability/player-stealth.h"
+#include "player-status/player-basic-statistics.h"
+#include "player-status/player-speed.h"
+#include "player-status/player-stealth.h"
 #include "player-status/player-hand-types.h"
 #include "player-status/player-infravision.h"
 #include "player/attack-defense-types.h"
