+++ src/player/player-status.cpp
@@ -60,12 +60,12 @@
 #include "object/object-mark-types.h"
 #include "perception/object-perception.h"
 #include "pet/pet-util.h"
-#include "player-ability/player-basic-statistics.h"
-#include "player-ability/player-speed.h"
-#include "player-ability/player-stealth.h"
 #include "player-info/avatar.h"
+#include "player-status/player-basic-statistics.h"
 #include "player-status/player-hand-types.h"
 #include "player-status/player-infravision.h"
+#include "player-status/player-speed.h"
+#include "player-status/player-stealth.h"
 #include "player/attack-defense-types.h"
 #include "player/digestion-processor.h"
 #include "player/mimic-info-table.h"
